+++ server/controllers/external/querytype.js
